Windows: Add support for UTF8 encoded filenames
This patch changes the way files will be created and opened on the
Windows platform.

Until now, filenames passed to functions such as H5Fcreate would work
only if they were encoded in the current codepage. This would not work
however, for example in the case when trying to open a filename
containing japanese characters on a system with the italian codepage.

A possible workaround would be through the use of GetShortPathName, but
this function does not work on all filesystems.

The only permanent solution involves the use of the wide character API,
as recommended by Microsoft (codepages are actively discouraged). For
this, the user needs to pass a Unicode UTF-8 filename to HDopen, which
is then fed to _wopen, after performing the conversion of the filename
to the corresponding wide character string (UTF-16).